Port Andy's tests over to Nick's implementation
