+++ andy_adapter.py
@@ -12,7 +12,7 @@ class Andy2048:
     def from_init_state(cls, init_state):
         andy_wrapper = cls()
         if isinstance(init_state, list) and len(init_state) == 16:
-            init_state =np.array(init_state).reshape((4,4))
+            init_state = np.array(init_state).reshape((4, 4))
         andy_wrapper.andy = BoardEnv.from_init_state(init_state)
         return andy_wrapper
 
@@ -39,4 +39,8 @@ class Andy2048:
         return self.board, self.score, self.andy.done
 
     def set_board(self, board):
-        self.andy.state = np.array(board).reshape(4,4)
+        self.andy.state = np.array(board).reshape(4, 4)
+
+    @classmethod
+    def random_direction(cls):
+        return BoardEnv.random_direction()
